+++ packages/react/src/components/subNavigation/tests/SubNavigation.spec.tsx
@@ -21,6 +21,19 @@ describe('SubNavigation', () => {
         ).not.toThrow();
 
         expect(() =>
+            shallow(
+                <SubNavigation
+                    items={[
+                        {id: 'a', label: 'A'},
+                        {id: 'b', label: 'B'},
+                    ]}
+                    selected="b"
+                    description="alphabet soup"
+                />
+            )
+        ).not.toThrow();
+
+        expect(() =>
             shallow(<SubNavigation items={[{id: 'a', label: 'A'}]} selected="not-existing-id" />)
         ).not.toThrow();
     });
@@ -52,10 +65,47 @@ describe('SubNavigation', () => {
             ],
         };
 
+        const propsWithDescription: ISubNavigationProps = {
+            items: [
+                {id: 'a', label: 'A'},
+                {id: 'b', label: 'B', description: 'Apple'},
+            ],
+        };
+
+        const propsWithCustomLabel: ISubNavigationProps = {
+            items: [
+                {id: 'a', label: <h1>Hello World</h1>},
+                {id: 'b', label: 'B', description: 'hello hello'},
+            ],
+        };
+
         it('should render one navigation link per item', () => {
             subNavigation = shallow(<SubNavigation {...basicProps} />);
 
             expect(subNavigation.find('li').length).toBe(basicProps.items.length);
+        });
+
+        it('should render a description', () => {
+            subNavigation = shallow(<SubNavigation {...propsWithDescription} />);
+            expect(subNavigation.find('div').length).toBe(3);
+            expect(subNavigation.find('div').last().hasClass('sub-navigation-item-description')).toBe(true);
+        });
+
+        it('should create items in div elements', () => {
+            subNavigation = shallow(<SubNavigation {...propsWithDescription} />);
+            expect(subNavigation.find('div').length).toBe(3);
+            expect(subNavigation.find('div').first().hasClass('sub-navigation-item-link-with-description-label')).toBe(
+                true
+            );
+        });
+
+        it('should respect the label when a description is supplied', () => {
+            subNavigation = shallow(<SubNavigation {...propsWithCustomLabel} />);
+            expect(subNavigation.find('div').length).toBe(2);
+            expect(
+                subNavigation.find('div').last().hasClass('sub-navigation-item-description body-m-book-subdued')
+            ).toBe(true);
+            expect(subNavigation.find('h1').length).toBe(1);
         });
 
         it('should have the "mod-selected" class on the selected item', () => {